backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
python-dotenv==1.0.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
